thông báo xác nhận và ui dashboard

+++ src/app/admin/blogs-category/trash/page.tsx
@@ -2,7 +2,7 @@
 
 import { Box, Typography, TextField, Select, MenuItem, InputLabel, FormControl, Button, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Checkbox, Chip, Tooltip, Stack, Pagination } from "@mui/material";
 import { MdDeleteOutline, MdOutlineSettingsBackupRestore } from "react-icons/md";
-
+import Alert from '@mui/material/Alert';
 import { useContext, useEffect, useState } from "react";
 import Link from "next/link";
 import { IoReturnDownBackOutline } from "react-icons/io5";
@@ -18,7 +18,8 @@ export default function BrandsTrashAdminPage() {
         currentPage: 1,
         blog_category: []
     });
-
+    const [alertMessage, setAlertMessage] = useState<string>("");
+    const [alertSeverity, setAlertSeverity] = useState<"success" | "error" | "info" | "warning">("info");
     const linkApi = 'https://freshskinweb.onrender.com/admin/blogs/category/trash';
 
     const [inputChecked, setInputChecked] = useState<number[]>([]);
@@ -153,8 +154,13 @@ export default function BrandsTrashAdminPage() {
 
         const dataResponse = await response.json();
 
-        if (dataResponse.code == 200) {
-            location.reload();
+        if (dataResponse.code === 200) {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("success");
+            setTimeout(() => location.reload(), 2000);
+        } else {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("error");
         }
     }
     // Hết Thay đổi trạng thái 1 sản phẩm
@@ -166,26 +172,35 @@ export default function BrandsTrashAdminPage() {
         const statusChange = changeMulti;
 
         if (statusChange == "delete-destroy") {
-            const path = `${linkApi}/delete`;
+            const confirm: boolean = window.confirm("Bạn có chắc muốn xóa vĩnh viễn những danh mục bài viết này không?");
+            if(confirm){
+                const path = `${linkApi}/delete`;
 
-            const data: any = {
-                id: inputChecked
+                const data: any = {
+                    id: inputChecked
+                }
+    
+                const response = await fetch(path, {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json"
+                    },
+                    body: JSON.stringify(data)
+                });
+    
+                const dataResponse = await response.json();
+    
+                if (dataResponse.code === 200) {
+                    setAlertMessage(dataResponse.message);
+                    setAlertSeverity("success");
+                    setTimeout(() => location.reload(), 2000);
+                } else {
+                    setAlertMessage(dataResponse.message);
+                    setAlertSeverity("error");
+                }
+    
             }
-
-            const response = await fetch(path, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify(data)
-            });
-
-            const dataResponse = await response.json();
-
-            if (dataResponse.code == 200) {
-                location.reload();
-            }
-
+            
             return;
         }
 
@@ -206,8 +221,13 @@ export default function BrandsTrashAdminPage() {
 
         const dataResponse = await response.json();
 
-        if (dataResponse.code == 200) {
-            location.reload();
+        if (dataResponse.code === 200) {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("success");
+            setTimeout(() => location.reload(), 2000);
+        } else {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("error");
         }
     }
 
@@ -222,39 +242,57 @@ export default function BrandsTrashAdminPage() {
 
     // Xóa vĩnh viễn một sản phẩm
     const handleDeleteOnebrand = async (id: number) => {
-        const path = `${linkApi}/delete/${id}`;
+        const confirm: boolean = window.confirm("Bạn có chắc muốn xóa vĩnh viễn danh mục bài viết này không?");
+        if(confirm){
+            const path = `${linkApi}/delete/${id}`;
 
-        const response = await fetch(path, {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json"
-            },
-        });
-
-        const dataResponse = await response.json();
-
-        if (dataResponse.code == 200) {
-            location.reload();
+            const response = await fetch(path, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+            });
+    
+            const dataResponse = await response.json();
+    
+            if (dataResponse.code === 200) {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("success");
+                setTimeout(() => location.reload(), 2000);
+            } else {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("error");
+            }
         }
+       
     }
     // Hết Xóa một sản phẩm
 
     // Khôi phục một sản phẩm
     const handleRestoreOnebrand = async (id: number) => {
-        const path = `${linkApi}/restore/${id}`;
+        const confirm: boolean = window.confirm("Bạn có chắc muốn khôi phục danh mục bài viết này không?");
+        if(confirm){
+            const path = `${linkApi}/restore/${id}`;
 
-        const response = await fetch(path, {
-            method: "PATCH",
-            headers: {
-                "Content-Type": "application/json"
-            },
-        });
-
-        const dataResponse = await response.json();
-
-        if (dataResponse.code == 200) {
-            location.reload();
+            const response = await fetch(path, {
+                method: "PATCH",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+            });
+    
+            const dataResponse = await response.json();
+    
+            if (dataResponse.code === 200) {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("success");
+                setTimeout(() => location.reload(), 2000);
+            } else {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("error");
+            }
         }
+       
     }
     // Hết Xóa một sản phẩm
