+++ src/app/admin/products-category/trash/page.tsx
@@ -2,7 +2,7 @@
 
 import { Box, Typography, TextField, Select, MenuItem, InputLabel, FormControl, Button, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Checkbox, Chip, Tooltip, Stack, Pagination } from "@mui/material";
 import { MdDeleteOutline, MdOutlineSettingsBackupRestore } from "react-icons/md";
-
+import Alert from '@mui/material/Alert';
 import { useContext, useEffect, useState } from "react";
 import Link from "next/link";
 import { IoReturnDownBackOutline } from "react-icons/io5";
@@ -19,7 +19,8 @@ export default function CategoriesTrashAdminPage() {
         currentPage: 1,
         product_category: []
     });
-
+    const [alertMessage, setAlertMessage] = useState<string>("");
+    const [alertSeverity, setAlertSeverity] = useState<"success" | "error" | "info" | "warning">("info");
     const linkApi = 'https://freshskinweb.onrender.com/admin/products/category/trash';
 
     const [inputChecked, setInputChecked] = useState<number[]>([]);
@@ -155,8 +156,13 @@ export default function CategoriesTrashAdminPage() {
 
         const dataResponse = await response.json();
 
-        if (dataResponse.code == 200) {
-            location.reload();
+        if (dataResponse.code === 200) {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("success");
+            setTimeout(() => location.reload(), 2000);
+        } else {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("error");
         }
     }
     // Hết Thay đổi trạng thái 1 sản phẩm
@@ -168,25 +174,33 @@ export default function CategoriesTrashAdminPage() {
         const statusChange = changeMulti;
 
         if (statusChange == "delete-destroy") {
-            const path = `${linkApi}/delete`;
-
-            const data: any = {
-                id: inputChecked
+            const confirm: boolean = window.confirm("Bạn có chắc muốn xóa vĩnh viễn những danh mục sản phẩm này không?");
+            if(confirm){
+                const path = `${linkApi}/delete`;
+                const data: any = {
+                    id: inputChecked
+                }
+    
+                const response = await fetch(path, {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json"
+                    },
+                    body: JSON.stringify(data)
+                });
+    
+                const dataResponse = await response.json();
+    
+                if (dataResponse.code === 200) {
+                    setAlertMessage(dataResponse.message);
+                    setAlertSeverity("success");
+                    setTimeout(() => location.reload(), 2000);
+                } else {
+                    setAlertMessage(dataResponse.message);
+                    setAlertSeverity("error");
+                }
             }
-
-            const response = await fetch(path, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify(data)
-            });
-
-            const dataResponse = await response.json();
-
-            if (dataResponse.code == 200) {
-                location.reload();
-            }
+           
 
             return;
         }
@@ -208,8 +222,13 @@ export default function CategoriesTrashAdminPage() {
 
         const dataResponse = await response.json();
 
-        if (dataResponse.code == 200) {
-            location.reload();
+        if (dataResponse.code === 200) {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("success");
+            setTimeout(() => location.reload(), 2000);
+        } else {
+            setAlertMessage(dataResponse.message);
+            setAlertSeverity("error");
         }
     }
 
@@ -224,39 +243,57 @@ export default function CategoriesTrashAdminPage() {
 
     // Xóa vĩnh viễn một sản phẩm
     const handleDeleteOnecategory = async (id: number) => {
-        const path = `${linkApi}/delete/${id}`;
+        const confirm: boolean = window.confirm("Bạn có chắc muốn xóa vĩnh viễn danh mục sản phẩm này không?");
+        if(confirm){
+            const path = `${linkApi}/delete/${id}`;
 
-        const response = await fetch(path, {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json"
-            },
-        });
-
-        const dataResponse = await response.json();
-
-        if (dataResponse.code == 200) {
-            location.reload();
+            const response = await fetch(path, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+            });
+    
+            const dataResponse = await response.json();
+    
+            if (dataResponse.code === 200) {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("success");
+                setTimeout(() => location.reload(), 2000);
+            } else {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("error");
+            }
         }
+       
     }
     // Hết Xóa một sản phẩm
 
     // Khôi phục một sản phẩm
     const handleRestoreOnecategory = async (id: number) => {
-        const path = `${linkApi}/restore/${id}`;
+        const confirm: boolean = window.confirm("Bạn có chắc muốn khôi phục danh mục sản phẩm này không?");
+        if(confirm){
+            const path = `${linkApi}/restore/${id}`;
 
-        const response = await fetch(path, {
-            method: "PATCH",
-            headers: {
-                "Content-Type": "application/json"
-            },
-        });
-
-        const dataResponse = await response.json();
-
-        if (dataResponse.code == 200) {
-            location.reload();
+            const response = await fetch(path, {
+                method: "PATCH",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+            });
+    
+            const dataResponse = await response.json();
+    
+            if (dataResponse.code === 200) {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("success");
+                setTimeout(() => location.reload(), 2000);
+            } else {
+                setAlertMessage(dataResponse.message);
+                setAlertSeverity("error");
+            }
         }
+       
     }
     // Hết Xóa một sản phẩm
 
@@ -296,6 +333,11 @@ export default function CategoriesTrashAdminPage() {
 
     return (
         <>
+         {alertMessage && (
+                                <Alert severity={alertSeverity} sx={{ mb: 2 }}>
+                                    {alertMessage}
+                                </Alert>
+                            )}
             {permissions?.includes("products-category_view") && permissions.includes("products-category_edit") && (
                 <Box p={3}>
                     {/* Header */}
